Add testing section to README
Sidekiq testing defaults to Sidekiq::Testing.inline!. In this mode the
sidekiq-rescue gem will not trigger. Testing the perform method will
still raise the error without rescuing it.

I thought it might be useful to add that information to the README, so
users of this gem have an easier time testing their jobs.

The solution I propose in the README is to disable the testing harness
in Sidekiq[1] and then to actually enqueue the job and run it.

[1] https://github.com/sidekiq/sidekiq/wiki/Testing#setup